% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_Data.R
\name{Binomial}
\alias{Binomial}
\title{Binomial R6 Class for Scale}
\description{
A logistic regression model using a GLM framework for the scale model.
}
\examples{
list2env(small_logistic_example, rlang::current_env())
model_glm <- glm(y ~ x - 1, family = binomial)
binomial_data <- Binomial$new(y = y, x = x)
}
\section{Super classes}{
\code{\link[Scale2:Data]{Scale2::Data}} -> \code{Scale2::GLM} -> \code{Binomial}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dispersion}}{The dispersion perameter of the Binomial Model}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Binomial-unit_deviance}{\code{Binomial$unit_deviance()}}
\item \href{#method-Binomial-v}{\code{Binomial$v()}}
\item \href{#method-Binomial-link}{\code{Binomial$link()}}
\item \href{#method-Binomial-grad_link}{\code{Binomial$grad_link()}}
\item \href{#method-Binomial-grad2_link}{\code{Binomial$grad2_link()}}
\item \href{#method-Binomial-logistic}{\code{Binomial$logistic()}}
\item \href{#method-Binomial-lap_ll}{\code{Binomial$lap_ll()}}
\item \href{#method-Binomial-grad_ll}{\code{Binomial$grad_ll()}}
\item \href{#method-Binomial-fisher_information}{\code{Binomial$fisher_information()}}
\item \href{#method-Binomial-new}{\code{Binomial$new()}}
\item \href{#method-Binomial-clone}{\code{Binomial$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="Data" data-id="alpha"><a href='../../Scale2/html/Data.html#method-Data-alpha'><code>Scale2::Data$alpha()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="Data" data-id="alpha_subsample"><a href='../../Scale2/html/Data.html#method-Data-alpha_subsample'><code>Scale2::Data$alpha_subsample()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="Data" data-id="div_alpha"><a href='../../Scale2/html/Data.html#method-Data-div_alpha'><code>Scale2::Data$div_alpha()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="Data" data-id="div_alpha_subsample"><a href='../../Scale2/html/Data.html#method-Data-div_alpha_subsample'><code>Scale2::Data$div_alpha_subsample()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="Data" data-id="phi"><a href='../../Scale2/html/Data.html#method-Data-phi'><code>Scale2::Data$phi()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="Data" data-id="phi_control_var"><a href='../../Scale2/html/Data.html#method-Data-phi_control_var'><code>Scale2::Data$phi_control_var()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="Data" data-id="phi_estimator"><a href='../../Scale2/html/Data.html#method-Data-phi_estimator'><code>Scale2::Data$phi_estimator()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="Data" data-id="phi_estimator_bounds"><a href='../../Scale2/html/Data.html#method-Data-phi_estimator_bounds'><code>Scale2::Data$phi_estimator_bounds()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="Data" data-id="precalc_control_var"><a href='../../Scale2/html/Data.html#method-Data-precalc_control_var'><code>Scale2::Data$precalc_control_var()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="Data" data-id="total_grad_ll"><a href='../../Scale2/html/Data.html#method-Data-total_grad_ll'><code>Scale2::Data$total_grad_ll()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="Data" data-id="total_lap_ll"><a href='../../Scale2/html/Data.html#method-Data-total_lap_ll'><code>Scale2::Data$total_lap_ll()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="Data" data-id="x_scale"><a href='../../Scale2/html/Data.html#method-Data-x_scale'><code>Scale2::Data$x_scale()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="Data" data-id="x_unscale"><a href='../../Scale2/html/Data.html#method-Data-x_unscale'><code>Scale2::Data$x_unscale()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="GLM" data-id="eta"><a href='../../Scale2/html/GLM.html#method-GLM-eta'><code>Scale2::GLM$eta()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="GLM" data-id="get_dispersion_param"><a href='../../Scale2/html/GLM.html#method-GLM-get_dispersion_param'><code>Scale2::GLM$get_dispersion_param()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="GLM" data-id="get_hessian_bound"><a href='../../Scale2/html/GLM.html#method-GLM-get_hessian_bound'><code>Scale2::GLM$get_hessian_bound()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="GLM" data-id="get_preconditioning_matrix"><a href='../../Scale2/html/GLM.html#method-GLM-get_preconditioning_matrix'><code>Scale2::GLM$get_preconditioning_matrix()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="GLM" data-id="log_pi_observed"><a href='../../Scale2/html/GLM.html#method-GLM-log_pi_observed'><code>Scale2::GLM$log_pi_observed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="GLM" data-id="log_pi_observed_i"><a href='../../Scale2/html/GLM.html#method-GLM-log_pi_observed_i'><code>Scale2::GLM$log_pi_observed_i()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Binomial-unit_deviance"></a>}}
\if{latex}{\out{\hypertarget{method-Binomial-unit_deviance}{}}}
\subsection{Method \code{unit_deviance()}}{
The unit deviance of a binomial model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Binomial$unit_deviance(mu, y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mu}}{Mean value mu}

\item{\code{y}}{Dependent variable y}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Binomial-v"></a>}}
\if{latex}{\out{\hypertarget{method-Binomial-v}{}}}
\subsection{Method \code{v()}}{
Variance Function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Binomial$v(mu)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mu}}{Mean value mu}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Binomial-link"></a>}}
\if{latex}{\out{\hypertarget{method-Binomial-link}{}}}
\subsection{Method \code{link()}}{
Link Function g(mu) = eta
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Binomial$link(mu)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mu}}{Mean value mu}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Binomial-grad_link"></a>}}
\if{latex}{\out{\hypertarget{method-Binomial-grad_link}{}}}
\subsection{Method \code{grad_link()}}{
Derivative of Link Function g(mu) = eta
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Binomial$grad_link(mu)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mu}}{Mean value mu}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Binomial-grad2_link"></a>}}
\if{latex}{\out{\hypertarget{method-Binomial-grad2_link}{}}}
\subsection{Method \code{grad2_link()}}{
Second derivative of Link Function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Binomial$grad2_link(mu)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mu}}{Mean value mu}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Binomial-logistic"></a>}}
\if{latex}{\out{\hypertarget{method-Binomial-logistic}{}}}
\subsection{Method \code{logistic()}}{
Logistic Function - Inverse of Link Function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Binomial$logistic(eta)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{eta}}{Eta. Inverse of value mu}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Binomial-lap_ll"></a>}}
\if{latex}{\out{\hypertarget{method-Binomial-lap_ll}{}}}
\subsection{Method \code{lap_ll()}}{
Laplacian of log likelihood at datum i
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Binomial$lap_ll(beta, i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{beta}}{- Unkown parameter $\beta$}

\item{\code{i}}{- Index of data matrix}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Binomial-grad_ll"></a>}}
\if{latex}{\out{\hypertarget{method-Binomial-grad_ll}{}}}
\subsection{Method \code{grad_ll()}}{
Gradient of log likelihood at datum i
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Binomial$grad_ll(beta, i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{beta}}{- Unkown parameter $\beta$}

\item{\code{i}}{- Index of data matrix}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Binomial-fisher_information"></a>}}
\if{latex}{\out{\hypertarget{method-Binomial-fisher_information}{}}}
\subsection{Method \code{fisher_information()}}{
Estimated Fisher Information for datum i
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Binomial$fisher_information(i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{- Index of data matrix}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Binomial-new"></a>}}
\if{latex}{\out{\hypertarget{method-Binomial-new}{}}}
\subsection{Method \code{new()}}{
Initialise Binomial Data Model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Binomial$new(y, x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y}}{Numeric vector length n. Dependent variables y}

\item{\code{x}}{Matrix of dimension n x d.  Response variables x}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Binomial-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Binomial-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Binomial$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
