% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cauchy_plots.R
\name{cauchy_density_path}
\alias{cauchy_density_path}
\title{Cauchy Kernel Density of Chain}
\usage{
cauchy_density_path(
  debug_hist,
  idx = seq_along(debug_hist),
  cauchy.x_0,
  cauchy.gamma
)
}
\arguments{
\item{debug_hist}{debug_hist list from SCALE Function}

\item{idx}{Index of mesh points to display}

\item{cauchy.x_0}{Location Parameter of Cauchy Target}

\item{cauchy.gamma}{Scale Parameter of Cauchy Target}
}
\description{
Cauchy Kernel Density of Chain
}
