% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{CauchyData}
\alias{CauchyData}
\title{R6 Cauchy Distribution}
\description{
Distribution Info for sampling from cauchy distribution, inherits from Data
Defines grad and laplacian with a flat prior
}
\section{Super class}{
\code{Scale2::Data} -> \code{CauchyData}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{mu}}{location parameter}

\item{\code{gamma}}{scale parameter}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CauchyData-pi}{\code{CauchyData$pi()}}
\item \href{#method-CauchyData-pi_ll}{\code{CauchyData$pi_ll()}}
\item \href{#method-CauchyData-grad_ll}{\code{CauchyData$grad_ll()}}
\item \href{#method-CauchyData-lap_ll}{\code{CauchyData$lap_ll()}}
\item \href{#method-CauchyData-new}{\code{CauchyData$new()}}
\item \href{#method-CauchyData-cauchy_scale}{\code{CauchyData$cauchy_scale()}}
\item \href{#method-CauchyData-phi_bounds_exact}{\code{CauchyData$phi_bounds_exact()}}
\item \href{#method-CauchyData-clone}{\code{CauchyData$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="Data" data-id="alpha"><a href='../../Scale2/html/Data.html#method-Data-alpha'><code>Scale2::Data$alpha()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="Data" data-id="alpha_subsample"><a href='../../Scale2/html/Data.html#method-Data-alpha_subsample'><code>Scale2::Data$alpha_subsample()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="Data" data-id="div_alpha"><a href='../../Scale2/html/Data.html#method-Data-div_alpha'><code>Scale2::Data$div_alpha()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="Data" data-id="div_alpha_subsample"><a href='../../Scale2/html/Data.html#method-Data-div_alpha_subsample'><code>Scale2::Data$div_alpha_subsample()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="Data" data-id="phi"><a href='../../Scale2/html/Data.html#method-Data-phi'><code>Scale2::Data$phi()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="Data" data-id="phi_control_var"><a href='../../Scale2/html/Data.html#method-Data-phi_control_var'><code>Scale2::Data$phi_control_var()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="Data" data-id="phi_estimator"><a href='../../Scale2/html/Data.html#method-Data-phi_estimator'><code>Scale2::Data$phi_estimator()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="Data" data-id="phi_estimator_bounds"><a href='../../Scale2/html/Data.html#method-Data-phi_estimator_bounds'><code>Scale2::Data$phi_estimator_bounds()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="Data" data-id="precalc_control_var"><a href='../../Scale2/html/Data.html#method-Data-precalc_control_var'><code>Scale2::Data$precalc_control_var()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="Data" data-id="total_grad_ll"><a href='../../Scale2/html/Data.html#method-Data-total_grad_ll'><code>Scale2::Data$total_grad_ll()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="Data" data-id="total_lap_ll"><a href='../../Scale2/html/Data.html#method-Data-total_lap_ll'><code>Scale2::Data$total_lap_ll()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="Data" data-id="x_scale"><a href='../../Scale2/html/Data.html#method-Data-x_scale'><code>Scale2::Data$x_scale()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="Data" data-id="x_unscale"><a href='../../Scale2/html/Data.html#method-Data-x_unscale'><code>Scale2::Data$x_unscale()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CauchyData-pi"></a>}}
\if{latex}{\out{\hypertarget{method-CauchyData-pi}{}}}
\subsection{Method \code{pi()}}{
Cauchy PDF
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CauchyData$pi(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Parameters of pdf}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CauchyData-pi_ll"></a>}}
\if{latex}{\out{\hypertarget{method-CauchyData-pi_ll}{}}}
\subsection{Method \code{pi_ll()}}{
Log Likelihood of Cauchy Distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CauchyData$pi_ll(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Parameters of ll}

\item{\code{i}}{not needed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CauchyData-grad_ll"></a>}}
\if{latex}{\out{\hypertarget{method-CauchyData-grad_ll}{}}}
\subsection{Method \code{grad_ll()}}{
Gradient of Log Likelihood of Cauchy Distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CauchyData$grad_ll(x, i = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Parameters of grad ll}

\item{\code{i}}{not needed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CauchyData-lap_ll"></a>}}
\if{latex}{\out{\hypertarget{method-CauchyData-lap_ll}{}}}
\subsection{Method \code{lap_ll()}}{
Laplacian of Log Likelihood of Cauchy Distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CauchyData$lap_ll(x, i = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Parameters of lap ll}

\item{\code{i}}{not needed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CauchyData-new"></a>}}
\if{latex}{\out{\hypertarget{method-CauchyData-new}{}}}
\subsection{Method \code{new()}}{
Cauchy Constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CauchyData$new(mu, gamma)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mu}}{Location Parameter}

\item{\code{gamma}}{Positive Scale Parameter}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CauchyData-cauchy_scale"></a>}}
\if{latex}{\out{\hypertarget{method-CauchyData-cauchy_scale}{}}}
\subsection{Method \code{cauchy_scale()}}{
Cauchy Scale Helper Function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CauchyData$cauchy_scale(x, mu, gamma)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Parameter}

\item{\code{mu}}{Location Parameter}

\item{\code{gamma}}{Positive Scale Parameter}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CauchyData-phi_bounds_exact"></a>}}
\if{latex}{\out{\hypertarget{method-CauchyData-phi_bounds_exact}{}}}
\subsection{Method \code{phi_bounds_exact()}}{
Exact bounds on phi
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CauchyData$phi_bounds_exact(x.l, x.u)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x.l}}{Lower Bound}

\item{\code{x.u}}{Upper Bounds}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CauchyData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CauchyData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CauchyData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
