% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{Data}
\alias{Data}
\title{R6 Base Data Distribution}
\description{
Parent of R6 Data Objects that implements killing rate and boundedness funtions
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Data-new}{\code{Data$new()}}
\item \href{#method-Data-total_grad_ll}{\code{Data$total_grad_ll()}}
\item \href{#method-Data-total_lap_ll}{\code{Data$total_lap_ll()}}
\item \href{#method-Data-phi}{\code{Data$phi()}}
\item \href{#method-Data-phi_control_var}{\code{Data$phi_control_var()}}
\item \href{#method-Data-phi_estimator}{\code{Data$phi_estimator()}}
\item \href{#method-Data-alpha}{\code{Data$alpha()}}
\item \href{#method-Data-div_alpha}{\code{Data$div_alpha()}}
\item \href{#method-Data-alpha_subsample}{\code{Data$alpha_subsample()}}
\item \href{#method-Data-div_alpha_subsample}{\code{Data$div_alpha_subsample()}}
\item \href{#method-Data-precalc_control_var}{\code{Data$precalc_control_var()}}
\item \href{#method-Data-phi_estimator_bounds}{\code{Data$phi_estimator_bounds()}}
\item \href{#method-Data-x_unscale}{\code{Data$x_unscale()}}
\item \href{#method-Data-x_scale}{\code{Data$x_scale()}}
\item \href{#method-Data-clone}{\code{Data$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Data-new"></a>}}
\if{latex}{\out{\hypertarget{method-Data-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Data$new(n, d, x_hat = NULL, hessian_bound = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Data-total_grad_ll"></a>}}
\if{latex}{\out{\hypertarget{method-Data-total_grad_ll}{}}}
\subsection{Method \code{total_grad_ll()}}{
Grad log likelihood over the full posterior returns an R^d vector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Data$total_grad_ll(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{parameter numeric length d}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Data-total_lap_ll"></a>}}
\if{latex}{\out{\hypertarget{method-Data-total_lap_ll}{}}}
\subsection{Method \code{total_lap_ll()}}{
Grad^2 log likelihood over the full posterior returns an R^d vector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Data$total_lap_ll(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{parameter numeric length d}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Data-phi"></a>}}
\if{latex}{\out{\hypertarget{method-Data-phi}{}}}
\subsection{Method \code{phi()}}{
Phi killing rate function over full posterior distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Data$phi(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{parameter numeric length d}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Data-phi_control_var"></a>}}
\if{latex}{\out{\hypertarget{method-Data-phi_control_var}{}}}
\subsection{Method \code{phi_control_var()}}{
Alternative phi function using x_hat
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Data$phi_control_var(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{parameter numeric length d}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Data-phi_estimator"></a>}}
\if{latex}{\out{\hypertarget{method-Data-phi_estimator}{}}}
\subsection{Method \code{phi_estimator()}}{
Phi Estimator sampling from datum indexes i, j
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Data$phi_estimator(x, i, j)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{parameter numeric length d}

\item{\code{i}}{Datum index i}

\item{\code{j}}{Datum index j}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Data-alpha"></a>}}
\if{latex}{\out{\hypertarget{method-Data-alpha}{}}}
\subsection{Method \code{alpha()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Data$alpha(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Data-div_alpha"></a>}}
\if{latex}{\out{\hypertarget{method-Data-div_alpha}{}}}
\subsection{Method \code{div_alpha()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Data$div_alpha(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Data-alpha_subsample"></a>}}
\if{latex}{\out{\hypertarget{method-Data-alpha_subsample}{}}}
\subsection{Method \code{alpha_subsample()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Data$alpha_subsample(x, i)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Data-div_alpha_subsample"></a>}}
\if{latex}{\out{\hypertarget{method-Data-div_alpha_subsample}{}}}
\subsection{Method \code{div_alpha_subsample()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Data$div_alpha_subsample(x, i)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Data-precalc_control_var"></a>}}
\if{latex}{\out{\hypertarget{method-Data-precalc_control_var}{}}}
\subsection{Method \code{precalc_control_var()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Data$precalc_control_var(x_hat = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Data-phi_estimator_bounds"></a>}}
\if{latex}{\out{\hypertarget{method-Data-phi_estimator_bounds}{}}}
\subsection{Method \code{phi_estimator_bounds()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Data$phi_estimator_bounds(maximal_distance)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Data-x_unscale"></a>}}
\if{latex}{\out{\hypertarget{method-Data-x_unscale}{}}}
\subsection{Method \code{x_unscale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Data$x_unscale(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Data-x_scale"></a>}}
\if{latex}{\out{\hypertarget{method-Data-x_scale}{}}}
\subsection{Method \code{x_scale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Data$x_scale(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Data-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Data-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Data$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
