% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_plots.R
\name{density_path}
\alias{density_path}
\title{Kernel Density Estimates Plot across mesh points}
\usage{
density_path(SCALE_info, idx = seq_along(debug_hist), unscale = TRUE, ...)
}
\arguments{
\item{SCALE_info}{Return object from SCALE}

\item{idx}{Mesh indexes to display}

\item{unscale}{Logical. Unscale particles}

\item{...}{Additional Parameters for ggplot}
}
\description{
Kernel Density Estimates Plot across mesh points
}
