% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_plots.R
\name{density_total}
\alias{density_total}
\title{Total Kernel Density Estimates Plot for ScaLE}
\usage{
density_total(SCALE_info, idx = seq_along(debug_hist), unscale = TRUE, ...)
}
\arguments{
\item{SCALE_info}{Return object from SCALE}

\item{idx}{Mesh indexes to include in total}

\item{unscale}{Logical. Unscale particles}

\item{...}{Additional Parameters for ggplot}
}
\description{
Total Kernel Density Estimates Plot for ScaLE
}
