% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_plots.R
\name{GLM_trace_path}
\alias{GLM_trace_path}
\title{Trace Path for GLM ScaLE Models}
\usage{
GLM_trace_path(
  SCALE_info,
  idx = seq_along(SCALE_info$debug_hist),
  unscale = TRUE
)
}
\arguments{
\item{SCALE_info}{SCALE return object}

\item{idx}{Index of mesh points to show}

\item{unscale}{Logical. Unscale Particles?}
}
\description{
Trace Path for GLM ScaLE Models
}
