% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_plots.R
\name{id_trace_path}
\alias{id_trace_path}
\title{Trace Path with ID's of Chain}
\usage{
id_trace_path(
  debug_hist,
  idx = seq_along(debug_hist),
  show_lines = TRUE,
  show_points = TRUE,
  sample_particles = TRUE,
  show_resample = TRUE
)
}
\arguments{
\item{debug_hist}{debug_hist of chain}

\item{idx}{Indexes of chain to show}

\item{show_lines}{Logical. Show Lines}

\item{show_points}{Logical. Show Points}

\item{sample_particles}{Logical. Show ID's}

\item{show_resample}{Logical. Show Resample Points}
}
\description{
Trace Path with ID's of Chain
}
