% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_Data.R
\name{Normal}
\alias{Normal}
\title{Normal GLM Model R6 Class}
\description{
A normal linear model class for specifying the distribution to the scale algorithm
}
\examples{
list2env(bivariate_normal_n20_b5neg2, rlang::current_env())
bi_normal <- Normal$new(y, x)
}
\section{Super classes}{
\code{\link[Scale2:Data]{Scale2::Data}} -> \code{Scale2::GLM} -> \code{Normal}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Normal-unit_deviance}{\code{Normal$unit_deviance()}}
\item \href{#method-Normal-v}{\code{Normal$v()}}
\item \href{#method-Normal-link}{\code{Normal$link()}}
\item \href{#method-Normal-grad_link}{\code{Normal$grad_link()}}
\item \href{#method-Normal-grad2_link}{\code{Normal$grad2_link()}}
\item \href{#method-Normal-logistic}{\code{Normal$logistic()}}
\item \href{#method-Normal-lap_ll}{\code{Normal$lap_ll()}}
\item \href{#method-Normal-new}{\code{Normal$new()}}
\item \href{#method-Normal-clone}{\code{Normal$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="Data" data-id="alpha"><a href='../../Scale2/html/Data.html#method-Data-alpha'><code>Scale2::Data$alpha()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="Data" data-id="alpha_subsample"><a href='../../Scale2/html/Data.html#method-Data-alpha_subsample'><code>Scale2::Data$alpha_subsample()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="Data" data-id="div_alpha"><a href='../../Scale2/html/Data.html#method-Data-div_alpha'><code>Scale2::Data$div_alpha()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="Data" data-id="div_alpha_subsample"><a href='../../Scale2/html/Data.html#method-Data-div_alpha_subsample'><code>Scale2::Data$div_alpha_subsample()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="Data" data-id="phi"><a href='../../Scale2/html/Data.html#method-Data-phi'><code>Scale2::Data$phi()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="Data" data-id="phi_control_var"><a href='../../Scale2/html/Data.html#method-Data-phi_control_var'><code>Scale2::Data$phi_control_var()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="Data" data-id="phi_estimator"><a href='../../Scale2/html/Data.html#method-Data-phi_estimator'><code>Scale2::Data$phi_estimator()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="Data" data-id="phi_estimator_bounds"><a href='../../Scale2/html/Data.html#method-Data-phi_estimator_bounds'><code>Scale2::Data$phi_estimator_bounds()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="Data" data-id="precalc_control_var"><a href='../../Scale2/html/Data.html#method-Data-precalc_control_var'><code>Scale2::Data$precalc_control_var()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="Data" data-id="total_grad_ll"><a href='../../Scale2/html/Data.html#method-Data-total_grad_ll'><code>Scale2::Data$total_grad_ll()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="Data" data-id="total_lap_ll"><a href='../../Scale2/html/Data.html#method-Data-total_lap_ll'><code>Scale2::Data$total_lap_ll()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="Data" data-id="x_scale"><a href='../../Scale2/html/Data.html#method-Data-x_scale'><code>Scale2::Data$x_scale()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="Data" data-id="x_unscale"><a href='../../Scale2/html/Data.html#method-Data-x_unscale'><code>Scale2::Data$x_unscale()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="GLM" data-id="eta"><a href='../../Scale2/html/GLM.html#method-GLM-eta'><code>Scale2::GLM$eta()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="GLM" data-id="fisher_information"><a href='../../Scale2/html/GLM.html#method-GLM-fisher_information'><code>Scale2::GLM$fisher_information()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="GLM" data-id="get_dispersion_param"><a href='../../Scale2/html/GLM.html#method-GLM-get_dispersion_param'><code>Scale2::GLM$get_dispersion_param()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="GLM" data-id="get_hessian_bound"><a href='../../Scale2/html/GLM.html#method-GLM-get_hessian_bound'><code>Scale2::GLM$get_hessian_bound()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="GLM" data-id="get_preconditioning_matrix"><a href='../../Scale2/html/GLM.html#method-GLM-get_preconditioning_matrix'><code>Scale2::GLM$get_preconditioning_matrix()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="GLM" data-id="grad_ll"><a href='../../Scale2/html/GLM.html#method-GLM-grad_ll'><code>Scale2::GLM$grad_ll()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="GLM" data-id="log_pi_observed"><a href='../../Scale2/html/GLM.html#method-GLM-log_pi_observed'><code>Scale2::GLM$log_pi_observed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Scale2" data-topic="GLM" data-id="log_pi_observed_i"><a href='../../Scale2/html/GLM.html#method-GLM-log_pi_observed_i'><code>Scale2::GLM$log_pi_observed_i()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Normal-unit_deviance"></a>}}
\if{latex}{\out{\hypertarget{method-Normal-unit_deviance}{}}}
\subsection{Method \code{unit_deviance()}}{
The unit deviance of a Normal model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Normal$unit_deviance(mu, y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mu}}{Mean value mu}

\item{\code{y}}{Dependent variable y}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Normal-v"></a>}}
\if{latex}{\out{\hypertarget{method-Normal-v}{}}}
\subsection{Method \code{v()}}{
Variance Function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Normal$v(mu)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mu}}{Mean value mu}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Normal-link"></a>}}
\if{latex}{\out{\hypertarget{method-Normal-link}{}}}
\subsection{Method \code{link()}}{
Link Function g(mu) = eta
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Normal$link(mu)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mu}}{Mean value mu}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Normal-grad_link"></a>}}
\if{latex}{\out{\hypertarget{method-Normal-grad_link}{}}}
\subsection{Method \code{grad_link()}}{
Derivative of Link Function g(mu) = eta
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Normal$grad_link(mu)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mu}}{Mean value mu}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Normal-grad2_link"></a>}}
\if{latex}{\out{\hypertarget{method-Normal-grad2_link}{}}}
\subsection{Method \code{grad2_link()}}{
Second derivative of Link Function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Normal$grad2_link(mu)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mu}}{Mean value mu}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Normal-logistic"></a>}}
\if{latex}{\out{\hypertarget{method-Normal-logistic}{}}}
\subsection{Method \code{logistic()}}{
Logistic Function - Inverse of Link Function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Normal$logistic(eta)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{eta}}{Eta. Inverse of value mu}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Normal-lap_ll"></a>}}
\if{latex}{\out{\hypertarget{method-Normal-lap_ll}{}}}
\subsection{Method \code{lap_ll()}}{
Laplacian of log likelihood at datum i
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Normal$lap_ll(beta, i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{beta}}{- Unkown parameter $\beta$}

\item{\code{i}}{- Index of data matrix}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Normal-new"></a>}}
\if{latex}{\out{\hypertarget{method-Normal-new}{}}}
\subsection{Method \code{new()}}{
Initialise Normal Data Model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Normal$new(y, x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y}}{Numeric vector length n. Dependent variables y}

\item{\code{x}}{Matrix of dimension n x d.  Response variables x}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Normal-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Normal-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Normal$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
