% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_plots.R
\name{plot_ess}
\alias{plot_ess}
\title{ESS Plot of ScaLE}
\usage{
plot_ess(SCALE_info, show_ess_thresh = FALSE)
}
\arguments{
\item{SCALE_info}{ScaLE return object}

\item{show_ess_thresh}{Logical. Show ESS threshold}
}
\description{
ESS Plot of ScaLE
}
