% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scale2-package.R
\docType{package}
\name{Scale2-package}
\alias{Scale2}
\alias{Scale2-package}
\title{Scale2: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ctrlbadger/Scale2}
  \item Report bugs at \url{https://github.com/ctrlbadger/Scale2/issues}
}

}
\keyword{internal}
