% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_plots.R
\name{trace_path}
\alias{trace_path}
\title{Trace Path of ScaLE}
\usage{
trace_path(debug_hist, idx = seq_along(debug_hist))
}
\arguments{
\item{debug_hist}{debug_hist list from SCALE return}

\item{idx}{Index vector of chains to show}
}
\description{
Trace Path of ScaLE
}
